+++ src/azure-firewall/azext_firewall/tests/latest/test_azure_firewall_scenario.py
@@ -33,11 +33,6 @@ class AzureFirewallScenario(ScenarioTest):
             self.check('threatIntelMode', 'Alert'),
             self.check('"Network.FTP.AllowActiveFTP"', 'true')
         ])
-        # validate same firewall name
-        from azure.cli.core.azclierror import ValidationError
-        with self.assertRaisesRegex(ValidationError, 'The specified firewall: af1 already exists.'):
-            self.cmd('network firewall create -g {rg} -n {af} --threat-intel-mode Alert --allow-active-ftp')
-
         self.cmd('network firewall update -g {rg} -n {af} --threat-intel-mode Deny --allow-active-ftp false', checks=[
             self.check('threatIntelMode', 'Deny'),
             self.not_exists('"Network.FTP.AllowActiveFTP"')
@@ -45,6 +40,31 @@ class AzureFirewallScenario(ScenarioTest):
         self.cmd('network firewall show -g {rg} -n {af}')
         self.cmd('network firewall list -g {rg}')
         self.cmd('network firewall delete -g {rg} -n {af}')
+
+    @ResourceGroupPreparer(name_prefix="cli_test_firewall_on_exist_", location="westus")
+    def test_firewall_on_exist(self):
+        self.kwargs.update({
+            "firewall_name": self.create_random_name("firewall-", 16),
+        })
+        self.cmd("network firewall create -n {firewall_name} -g {rg} --threat-intel-mode Alert")
+
+        from azure.cli.core.azclierror import ValidationError
+        with self.assertRaisesRegex(ValidationError, f"The specified firewall: {self.kwargs['firewall_name']} already exists."):
+            self.cmd("network firewall create -n {firewall_name} -g {rg} --threat-intel-mode Deny --on-exist error")
+        self.cmd(
+            "network firewall create -n {firewall_name} -g {rg} --threat-intel-mode Deny --on-exist skip",
+            checks=[
+                self.check("threatIntelMode", "Alert")
+            ]
+        )
+        self.cmd(
+            "network firewall create -n {firewall_name} -g {rg} --threat-intel-mode Deny --on-exist overwrite",
+            checks=[
+                self.check("threatIntelMode", "Deny")
+            ]
+        )
+
+        self.cmd("network firewall delete -n {firewall_name} -g {rg}")
 
     @ResourceGroupPreparer(name_prefix="cli_test_firewall_with_additional_log_", location="westus")
     def test_firewall_with_additional_log(self):